More changes for Boost 1.45
git-svn-id: http://svn.boost.org/svn/boost/trunk@64985 b8fc166d-592f-0410-95f2-cb63ce0dd405

+++ boost/interprocess/detail/os_file_functions.hpp
@@ -127,7 +127,7 @@ inline bool truncate_file (file_handle_t hnd, std::size_t size)
    if(!winapi::get_file_size(hnd, filesize))
       return false;
 
-   if(size > filesize){
+   if(size > (unsigned long long)filesize){
       if(!winapi::set_file_pointer_ex(hnd, filesize, 0, winapi::file_begin)){
          return false;
       }
